blacklist check in get.sh

+++ get.sh
@@ -16,17 +16,21 @@ while true; do
     artist=$(playerctl metadata --format '{{ artist }}' 2>/dev/null)
     title=$(playerctl metadata --format '{{ title }}' 2>/dev/null)
 
-    if [[ -n "$artist" && -n "$title" ]]; then
-        echo "$artist - $title" > "$base_dir/playing.txt"
-    elif [[ -n "$artist" ]]; then
-        echo "$artist" > "$base_dir/playing.txt"
-    elif [[ -n "$title" ]]; then
-        echo "$title" > "$base_dir/playing.txt"
-    else
+    if grep -iq -f "$base_dir/blacklist" <<< "$artist $title"; then
         > "$base_dir/playing.txt"
-    fi
-    if [[ $(grep -o '-' "$base_dir/playing.txt" | wc -l) -ge 2 ]]; then
-        sed -i 's/^[^-]*- *//' "$base_dir/playing.txt"
+    else
+        if [[ -n "$artist" && -n "$title" ]]; then
+            echo "$artist - $title" > "$base_dir/playing.txt"
+        elif [[ -n "$artist" ]]; then
+            echo "$artist" > "$base_dir/playing.txt"
+        elif [[ -n "$title" ]]; then
+            echo "$title" > "$base_dir/playing.txt"
+        else
+            > "$base_dir/playing.txt"
+        fi
+        if [[ $(grep -o '-' "$base_dir/playing.txt" | wc -l) -ge 2 ]]; then
+            sed -i 's/^[^-]*- *//' "$base_dir/playing.txt"
+        fi
     fi
 
     sleep 5
